refactor(daemon): split the web-tab URL algebra out of webtab_proxy.go (#1145)
webtab_proxy.go crossed the 1000-line lint limit at 1125: #1870 added the
escaped-path normalizer and network-path guard, #1873 added the socket
transport, and this PR added the remainder/query helpers. Splitting it, not
grandfathering it — the file-length lint exists to force exactly this.

webtab_url.go takes the PURE layer: the mirror's URL algebra (split, prefix,
rewrite of paths and queries), with no handler, manager, or HTTP in it. That is
also the layer with a single shared invariant worth stating once — a URL's
decoded and escaped views must never be composed side by side, because net/url
drops RawPath the moment they drift — so the functions that must obey it now sit
together under it.

Pure code motion: no behavior change, no signature change, tests untouched.
webtab_proxy.go 1125 -> 773; webtab_url.go 371.
